backend/requirements.txt
anthropic==0.42.0
aiosqlite==0.20.0
fastapi==0.115.6
google-generativeai==0.8.3
gunicorn==23.0.0
httpx==0.28.1
mlflow==2.19.0
openai==1.58.1
pydantic==2.10.4
pydantic-settings==2.7.0
pytest==8.3.4
python-dotenv==1.0.1
PyYAML==6.0.2
SQLAlchemy==2.0.37
slowapi==0.1.9
structlog==24.4.0
uvicorn==0.32.1